chore: database 환경 변수 추가
