--- api/src/main/java/com/tago/api/auth/dto/request/TokenReissueRequest.java
@@ -1,5 +1,0 @@
-package com.tago.api.auth.dto.request;
-
-public class TokenReissueRequest {
-    private String refreshToken;
-}
