+++ api/src/main/resources/application.yml
@@ -1,8 +1,8 @@
 spring:
   datasource:
     url: jdbc:mysql://localhost:3306/tago?serverTimezone=Asia/Seoul
-    username: root
-    password: 2532
+    username: ${DATABASE_USERNAME}
+    password: ${DATABASE_PASSWORD}
     driver-class-name: com.mysql.cj.jdbc.Driver
   sql:
     init:
@@ -20,7 +20,7 @@ spring:
     defer-datasource-initialization: true
 
 jwt:
-  secret-key: ${JWT_SECRET_KEY}
+  secretKey: ${JWT_SECRET_KEY}
 
 oauth:
   kakao:
